+++ packages/core-gatsby-ui/src/navigation/aside.tsx
@@ -25,20 +25,23 @@ export const Aside: React.FC<Props> = (props) => {
   return (
     <BoxV2 as={'aside'} className={styles.wrapper}>
       <Stack as={'nav'} className={styles.wrapper}>
-        {props.items?.filter(filterItemDepthPredicate).map((item) => {
-          const href = `#${formatAnchorId(item?.value)}`;
-          const selected = formatAnchorId(item?.value) === currentId;
-          return (
-            <a href={href} className={styles.link} key={formatAnchorId(item?.value)}>
-              <Label
-                id={`aside-link-${item?.value}`}
-                className={`${styles.link} ${selected ? styles.linkSelected : ''}`}
-              >
-                {item?.value}
-              </Label>
-            </a>
-          );
-        })}
+        <Label>Contents</Label>
+        <div className={styles.linksWrapper}>
+          {props.items?.filter(filterItemDepthPredicate).map((item) => {
+            const href = `#${formatAnchorId(item?.value)}`;
+            const selected = formatAnchorId(item?.value) === currentId;
+            return (
+              <a href={href} className={styles.link} key={formatAnchorId(item?.value)}>
+                <Label
+                  id={`aside-link-${item?.value}`}
+                  className={`${styles.link} ${selected ? styles.linkSelected : ''}`}
+                >
+                  {item?.value}
+                </Label>
+              </a>
+            );
+          })}
+        </div>
       </Stack>
     </BoxV2>
   );