+++ packages/core-gatsby-ui/src/sections/section-divider.treat.ts
@@ -10,7 +10,7 @@ export const styles = {
     '@media': {
       [cssTheme.layout.media.tablet]: {
         maxWidth: '80ch',
-        padding: `${cssTheme.sizing.var.x7} 0 ${cssTheme.sizing.var.x7} 0`,
+        padding: `${cssTheme.sizing.var.x7} ${cssTheme.layout.var.contentMargins}`,
       },
     },
   })),
@@ -19,5 +19,10 @@ export const styles = {
     height: `100%`,
     minHeight: `50vh`,
     maxWidth: '100%',
+    '@media': {
+      [cssTheme.layout.media.tablet]: {
+        width: '50vw',
+      },
+    },
   })),
 };