+++ packages/core-react-ui/src/layout/stack.tsx
@@ -10,27 +10,27 @@ type Props = CommonComponentProps &
     gap: GapProp;
   }>;
 
-export const Stack = React.memo(
-  React.forwardRef<any, Props>(function Stack({ id, style, className = '', as, gap = ['0px'], ...props } = {}, ref) {
-    const { styles } = useStyles(styleRefs);
-    const [mobileGap, tabletGap, desktopGap] = gap;
+export const Stack = React.forwardRef<any, Props>(function Stack(
+  { id, style, className = '', as, gap = ['0px'], ...props } = {},
+  ref
+) {
+  const { styles } = useStyles(styleRefs);
+  const [mobileGap, tabletGap, desktopGap] = gap;
 
-    return React.createElement(as || 'div', {
-      id,
-      className: `${className || ''} ${styles.wrapper}`,
-      style: {
-        ...style,
-        // @ts-ignore
-        '--mobile-gap': mobileGap,
-        '--tablet-gap': tabletGap || mobileGap,
-        '--desktop-gap': desktopGap || tabletGap || mobileGap,
-      },
-      ...props,
-      ref,
-    });
-  })
-  // areEqual
-);
+  return React.createElement(as || 'div', {
+    id,
+    className: `${className || ''} ${styles.wrapper}`,
+    style: {
+      ...style,
+      // @ts-ignore
+      '--mobile-gap': mobileGap,
+      '--tablet-gap': tabletGap || mobileGap,
+      '--desktop-gap': desktopGap || tabletGap || mobileGap,
+    },
+    ref,
+    ...props,
+  });
+});
 
 function areEqual(prevProps: Props, nextProps: Props) {
   const previousGap = prevProps.gap;