+++ packages/core-react-ui-app/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/newrade/newrade/compare/@newrade/core-react-ui-app@1.6.0...@newrade/core-react-ui-app@1.6.1) (2021-03-10)
+
+**Note:** Version bump only for package @newrade/core-react-ui-app
+
 # [1.6.0](https://github.com/newrade/newrade/compare/@newrade/core-react-ui-app@1.5.9...@newrade/core-react-ui-app@1.6.0) (2021-03-09)
 
 ### Bug Fixes